pyproject.toml
[project]
name = "quickpickbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "redis",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.40",
]
